+++ tests/testthat/test-connect.R
@@ -1,7 +1,7 @@
 context("test Connect R6 class")
 
 test_that("preserves provided values", {
-  host <- "myhost.example.com"
+  host <- "http://myhost.example.com"
   api_key <- "fake"
 
   con <- Connect$new(host = host, api_key = api_key)
@@ -11,10 +11,24 @@ test_that("preserves provided values", {
 })
 
 test_that("trailing slash removed from host", {
-  con <- Connect$new(host = "myhost.example.com/", api_key = "fake")
-  con2 <- Connect$new(host = "myhost.example.com", api_key = "fake")
+  con <- Connect$new(host = "http://myhost.example.com/", api_key = "fake")
+  con2 <- Connect$new(host = "http://myhost.example.com", api_key = "fake")
 
   expect_identical(con$host, con2$host)
+})
+
+test_that("error if protocol not defined", {
+  con <- Connect$new(host = "https://myhost.example.com", api_key = "fake")
+
+  expect_error(
+    Connect$new("test.example.com", "fake"),
+    "protocol"
+  )
+
+  expect_error(
+    Connect$new("://test.example.com", "fake"),
+    "protocol"
+  )
 })